+++ NEWS.md
@@ -12,6 +12,10 @@
 - new function `group_data()` to extract the grouping structure (#3489).
 - new function `group_rows()` to get a list of row indices for each group (#3489).
 - `first()` and `last()` hybrid functions fall back to R evaluation when given no arguments (#3589). 
+- new function `rows()` to get a list of row indices for each group (#3489).
+- new method `rows()` to get a list of row indices for each group (#3489).
+- new method `group_data()` (#3489).
+- joins no longer make lazy grouped data (#3566). 
 
 # dplyr 0.7.5.9001